Renamed model package's packages to adhere to conventions
