+++ Phonebook/src/main/java/application/main/controller/AddressController.java
@@ -1,6 +1,6 @@
 package application.main.controller;
 
-import application.main.model.Entity.Address;
+import application.main.model.entity.Address;
 import application.main.service.Dispatcher;
 import application.main.service.Interfaces.IDispatcher;
 import org.springframework.http.HttpStatus;
