+++ Phonebook/src/main/java/application/main/model/Database/DAOs/PersonDAO.java
@@ -2,8 +2,8 @@ package application.main.model.Database.DAOs;
 
 import application.main.model.Database.DatabaseHandlerFactory;
 import application.main.model.Database.Interfaces.IPersonDAO;
-import application.main.model.Entity.DTOs.SimplePersonDTO;
-import application.main.model.Entity.Person;
+import application.main.model.entity.dto.SimplePersonDTO;
+import application.main.model.entity.Person;
 import com.microsoft.sqlserver.jdbc.SQLServerDriver;
 
 import java.sql.*;
