+++ Phonebook/src/main/java/application/main/model/Database/Interfaces/IAddressDAO.java
@@ -1,6 +1,6 @@
 package application.main.model.Database.Interfaces;
 
-import application.main.model.Entity.Address;
+import application.main.model.entity.Address;
 
 import java.util.ArrayList;
