+++ Phonebook/src/main/java/application/main/service/ContactService.java
@@ -2,7 +2,7 @@ package application.main.service;
 
 import application.main.model.Database.DAOs.ContactInfoDAO;
 import application.main.model.Database.Interfaces.IContactInfoDAO;
-import application.main.model.Entity.ContactInfo;
+import application.main.model.entity.ContactInfo;
 import application.main.service.interfaces.IContactService;
 
 import java.util.List;
