+++ Phonebook/src/main/java/application/main/service/interfaces/IAddressService.java
@@ -1,6 +1,6 @@
 package application.main.service.interfaces;
 
-import application.main.model.Entity.Address;
+import application.main.model.entity.Address;
 
 import java.util.List;
