+++ Phonebook/src/test/java/AddressTest.java
@@ -1,5 +1,5 @@
-import application.main.model.Entity.Address;
-import application.main.model.Entity.ContactInfo;
+import application.main.model.entity.Address;
+import application.main.model.entity.ContactInfo;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
